hardware/requirements.txt
RPi.GPIO
scapy
requests